enumerator invalidation should never throw
Even if this is compiled with checked arithmetic as the default,
wrapping around is the expected behavior.

With a 64-bit integer, this is not expected to occur, though perhaps it
could in very long-running programs performing numerous interleaved
insertions and deletions.

Also, I'm not sure _version really should be 64-bit. It might be changed
to 32-bit, in which case wrapping is a more plausible possibility.