added terms of use
Using the Wolfram|Alpha API (directly, as a client) requires that
software have a Terms of Use that contains particular information,
including some particular text.

I am unsure whether it is permitted to call it something other than a
Terms of Use, so I figure, better to be safe, at least until I have
confirmation otherwise.

As long as the master branch doesn't contain code to use the
Wolfram|Alpha API, I don't intend the terms of use comments added in
this commit to appear there.

The version of this code that does not access the Wolfram|Alpha API
should not be restricted in this way, even if the restriction would have
minimal or no effects on the user in the absence of the ability to
access the Wolfram|Alpha API using that code. I might end up releasing
that code as public domain or under a FOSS license, so I don't want to
impose these terms on its use.

As for whether or not the version that does contain code to access the
Wolfram|Alpha API can be released in free-as-in-freedom manner, I am
unsure. If accessing that API proves useful and the code to do so is
retained (i.e., if this branch and its descendants don't merely whither
away) then I'll want to look into that.

In particular, there are libraries such as
https://github.com/Genbox/WolframAlphaNet that are released under FOSS
licenses (in the case of that library, the MIT license) and that
themselves contain code that accesses the Wolfram|Alpha API. They are
for use by applications that have an API key.

The situation with the code in this repository _might_ be similar, in
that, at least as it currently stands, this is not a weapp, and the only
ways to run it are either (a) to have a local copy, including one's own
Wolfram|Alpha API key if one wishes to use that functionality, or (b)
use some other software into which some or all of this code has been
integrated.