security: bump RestSharp from 106.6.7 to 106.15.0
This fixes CVE-2021-27293. For details, see:

    https://github.com/advisories/GHSA-9pq7-rcxv-47vq

RestSharp 106.6.7 was resolved as the minimum version required to
satisfy the indirect dependency through WolframAlphaNET, so I have
made it a direct dependency and pinned it at the most recent patch
version in 106.6. (I've also pinned WolframAlphaNET at 2.0.0, which
is the same version as before.)